NEW: ContentTypeCacheFixMixin for TestCases
TestCase mixin to fill the ContentType cache to avoid flaky database queries.

The Problem is, that the ContentTypeManager used a shared cache for lookups.
If your tests checks the made database queries, then you may get different results
depending on how the cache is filled.

With this Mixin the ContentTypeManager cache is always filled and all tests will never
hit the databases...

Additional to this: Update README:

* Update "make" help
* Add a note about README update via tests and Python 3.9