Reimplement Playwright test helper
Add a new `PlaywrightTestCase` class for Playwright tests to replace the old
`PyTestPlaywrightBaseTestCase`

Leave the old test code and add deprecation warning to all old helper.
We should remove them in the future. Also remove the old tests (`PlaywrightTestHelperTestCase`) for
the old helper.

Add Firefox and Chromium tests for the existing "User Timezone" to check the used JavaScript.