Optimize FeatureFlag caching and bugfix FeatureFlagTestCaseMixin
Before this PR the state of a FeatureFlag was only stored into the cache on state change.

The result is, that every `.is_enabled` call will hit the database until the state was change for
the first time.

Change this and store the state to cache + database on first usage.

Also set a per-process and thread-safe cache backend for tests. This is needed if tests runs with
`--parallel` add.
Ass helper `get_feature_flag_cache_info()` and `get_feature_flag_db_info()` Think they are only
needed in own tests, so i didn't add a DocString to hide them from README.