Enhance translation helper
raise `ValidationError` if unknown translations should be stored. For normal usage this should never
happen, because the user get's a form only with known languages.
Before this PR the `_reduce_to_known_languages()` is a kind of suppressed error: It just removed all
"unknown" lanuages, but totally silent. Because of this, I spent a long time looking for a bug in my
own code. But I had only used an unsupported language in the test.

Bind '<UNTRANSLATED>' string to `FieldTranslation.UNTRANSLATED` so it can be used from external code
to compare.

Add a `FieldTranslation.__repr__` and make it helpful for compare in tests.

Add more tests and set `unittest.util._MAX_LENGTH` for better test output.