Add Django LogEntry for changing feature flag states.
Additional changes:

`HtmlAssertionMixin.assert_messages()` will not clear the existing messages. Implement this is not
really easy :( As a work-a-round: Implement `get_fresh_client()` to "clear" all messages by create a
new session. But keep the current user logged-in.

Also enhance `assert_messages()` and display a clean copy&paste code block for updating the test
code.

And add `LogEntryAdmin` to see the created LogEntries.

Remove obsolete Django 3.2 snapshot files and update all other, because of line ending changes in
https://github.com/boxine/bx_py_utils/commit/86687ad43aef4baa71fab5aef2969ada006ef1ad