Bugfix create or update if UUID model field updated with a UUID string.
Django/database will convert a UUID string automatically to a UUID object.

But the `update_model_field_callback()` will just compare the objects
and this will be different if `create_or_update2()` will be called with a UUID string.

So convert the database UUID value to string, so that `update_model_field_callback()` detect and
"unchange" or "changed" value correctly.