Add retval_factory parameter to @if_enabled
Currently, if a feature flag is disabled, the @if_enabled decorator will always return `None`. This may break the caller if it expects a return value of the same type as it would receive when the flag is enabled, or make the caller code more verbose because of necessary `None` checks. Adding the new optional `retval_factory` parameter lets the caller specify which non-`None` value should be returned in that case.