Remove cache layer from feature flags
There were 2 problems with the current implementation of feature flags:
1. They suggested being resilient to cache outages, but they would still attempt to write changed state to the cache, which would result in an unhandled exception.
2. After fixing 1, cache and database would remain in a split-brain state until either the related cache keys were cleared or the state would manually we changed. This is surprising (and in some cases dangerous) behavior.

There is no easy way to fix the split-brain state, so remove the caching layer all together and only store the state in the database. If the user does not want to query the database for each state evaluation he should opt for the recently introduced in-process state caching.